+++ packages/core/src/plugins/with-selection.ts
@@ -12,15 +12,30 @@ import {
     getSelectedElements,
     removeSelectedElement
 } from '../utils/selected-element';
-import { PlaitElement, PlaitPointerType, SELECTION_BORDER_COLOR, SELECTION_FILL_COLOR } from '../interfaces';
-import { getRectangleByElements } from '../utils/element';
-import { BOARD_TO_IS_SELECTION_MOVING, BOARD_TO_TEMPORARY_ELEMENTS } from '../utils/weak-maps';
-import { ACTIVE_STROKE_WIDTH, ATTACHED_ELEMENT_CLASS_NAME, SELECTION_RECTANGLE_CLASS_NAME } from '../constants/selection';
-import { drawRectangle, isDragging, preventTouchMove, setDragging, throttleRAF, toHostPoint, toViewBoxPoint } from '../utils';
+import { PlaitElement, PlaitGroup, PlaitPointerType, SELECTION_BORDER_COLOR, SELECTION_FILL_COLOR } from '../interfaces';
+import { ATTACHED_ELEMENT_CLASS_NAME } from '../constants/selection';
+import { drawRectangle, isDragging, preventTouchMove, throttleRAF, toHostPoint, toViewBoxPoint } from '../utils';
 import { PlaitOptionsBoard, PlaitPluginOptions } from './with-options';
 import { PlaitPluginKey } from '../interfaces/plugin-key';
 import { Selection } from '../interfaces/selection';
 import { PRESS_AND_MOVE_BUFFER } from '../constants';
+import {
+    getElementsByGroup,
+    getGroupByElement,
+    getSelectedGroups,
+    isSelectedElementsIncludeGroup,
+    getElementsInGroupByElement
+} from '../utils/group';
+import {
+    clearSelectionMoving,
+    createSelectionRectangleG,
+    deleteTemporaryElements,
+    getTemporaryElements,
+    isHandleSelection,
+    isSelectionMoving,
+    isSetSelectionOperation,
+    setSelectionMoving
+} from '../utils/selection';
 
 export interface WithPluginOptions extends PlaitPluginOptions {
     isMultiple: boolean;
@@ -73,7 +88,7 @@ export function withSelection(board: PlaitBoard) {
             // prevent text from being selected
             event.preventDefault();
         }
-        if (start && PlaitBoard.isPointer(board, PlaitPointerType.selection)) {
+        if (PlaitBoard.isPointer(board, PlaitPointerType.selection) && start) {
             const movedTarget = toViewBoxPoint(board, toHostPoint(board, event.x, event.y));
             const rectangle = RectangleClient.getRectangleByPoints([start, movedTarget]);
             selectionMovingG?.remove();
@@ -129,7 +144,6 @@ export function withSelection(board: PlaitBoard) {
                 Transforms.setSelection(board, null);
             }
         }
-
         start = null;
         end = null;
         isTextSelection = false;
@@ -158,6 +172,7 @@ export function withSelection(board: PlaitBoard) {
                 if (!options.isMultiple && elements.length > 1) {
                     elements = [elements[0]];
                 }
+                const hitElementsWithGroup = elements.some(item => item.groupId);
                 if (isShift) {
                     if (board.selection && Selection.isCollapsed(board.selection)) {
                         const newSelectedElements = [...getSelectedElements(board)];
@@ -179,7 +194,30 @@ export function withSelection(board: PlaitBoard) {
                         cacheSelectedElements(board, newSelectedElements);
                     }
                 } else {
-                    const newSelectedElements = [...elements];
+                    let newSelectedElements = [...elements];
+                    if (hitElementsWithGroup) {
+                        const isCollapsed = Selection.isCollapsed(board.selection!);
+                        if (!isCollapsed) {
+                            newSelectedElements = [];
+                            elements.forEach(item => {
+                                if (!item.groupId) {
+                                    newSelectedElements.push(item);
+                                } else {
+                                    newSelectedElements.push(...getElementsInGroupByElement(board, item));
+                                }
+                            });
+                        } else {
+                            const groups = getGroupByElement(board, elements[0], true) as PlaitGroup[];
+                            const selectedGroups = getSelectedGroups(board, groups);
+                            if (isSelectedElementsIncludeGroup(selectedGroups)) {
+                                if (selectedGroups.length > 1) {
+                                    newSelectedElements = getElementsByGroup(board, selectedGroups[selectedGroups.length - 2], true);
+                                }
+                            } else {
+                                newSelectedElements = getElementsByGroup(board, groups[groups.length - 1], true);
+                            }
+                        }
+                    }
                     cacheSelectedElements(board, newSelectedElements);
                 }
                 const newElements = getSelectedElements(board);
@@ -202,8 +240,9 @@ export function withSelection(board: PlaitBoard) {
                 const currentSelectedElements = getSelectedElements(board);
                 if (currentSelectedElements.length && currentSelectedElements.length > 1) {
                     if (
-                        currentSelectedElements.length !== previousSelectedElements.length ||
-                        currentSelectedElements.some((c, index) => c !== previousSelectedElements[index])
+                        previousSelectedElements &&
+                        (currentSelectedElements.length !== previousSelectedElements.length ||
+                            currentSelectedElements.some((c, index) => c !== previousSelectedElements[index]))
                     ) {
                         selectionRectangleG?.remove();
                         selectionRectangleG = createSelectionRectangleG(board);
@@ -225,62 +264,4 @@ export function withSelection(board: PlaitBoard) {
     });
 
     return board;
-}
-
-export function isHandleSelection(board: PlaitBoard) {
-    const options = (board as PlaitOptionsBoard).getPluginOptions<WithPluginOptions>(PlaitPluginKey.withSelection);
-    return board.pointer !== PlaitPointerType.hand && !options.isDisabledSelect && !PlaitBoard.isReadonly(board);
-}
-
-export function isSetSelectionOperation(board: PlaitBoard) {
-    return !!board.operations.find(value => value.type === 'set_selection');
-}
-
-export function getTemporaryElements(board: PlaitBoard) {
-    const ref = BOARD_TO_TEMPORARY_ELEMENTS.get(board);
-    if (ref) {
-        return ref.elements;
-    } else {
-        return undefined;
-    }
-}
-
-export function getTemporaryRef(board: PlaitBoard) {
-    return BOARD_TO_TEMPORARY_ELEMENTS.get(board);
-}
-
-export function deleteTemporaryElements(board: PlaitBoard) {
-    BOARD_TO_TEMPORARY_ELEMENTS.delete(board);
-}
-
-export function isSelectionMoving(board: PlaitBoard) {
-    return !!BOARD_TO_IS_SELECTION_MOVING.get(board);
-}
-
-export function setSelectionMoving(board: PlaitBoard) {
-    PlaitBoard.getBoardContainer(board).classList.add('selection-moving');
-    BOARD_TO_IS_SELECTION_MOVING.set(board, true);
-    setDragging(board, true);
-}
-
-export function clearSelectionMoving(board: PlaitBoard) {
-    PlaitBoard.getBoardContainer(board).classList.remove('selection-moving');
-    BOARD_TO_IS_SELECTION_MOVING.delete(board);
-    setDragging(board, false);
-}
-
-export function createSelectionRectangleG(board: PlaitBoard) {
-    const elements = getSelectedElements(board);
-    const rectangle = getRectangleByElements(board, elements, false);
-    if (rectangle.width > 0 && rectangle.height > 0 && elements.length > 1) {
-        const selectionRectangleG = drawRectangle(board, RectangleClient.inflate(rectangle, ACTIVE_STROKE_WIDTH), {
-            stroke: SELECTION_BORDER_COLOR,
-            strokeWidth: ACTIVE_STROKE_WIDTH,
-            fillStyle: 'solid'
-        });
-        selectionRectangleG.classList.add(SELECTION_RECTANGLE_CLASS_NAME);
-        PlaitBoard.getElementActiveHost(board).append(selectionRectangleG);
-        return selectionRectangleG;
-    }
-    return null;
 }