+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # plait
 
+## 0.51.2
+
+### Patch Changes
+
+-   [`5679c8bd`](https://github.com/worktile/plait/commit/5679c8bd40461ca1339d334557a40930bd985689) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - add return to stop moving for special scene
+
 ## 0.51.1
 
 ## 0.51.0
@@ -31,7 +37,7 @@
 ### Patch Changes
 
 -   [#744](https://github.com/worktile/plait/pull/744) [`712ccb12`](https://github.com/worktile/plait/commit/712ccb12a6113c5b59eb48b022732824ff529044) Thanks [@huanhuanwa](https://github.com/huanhuanwa)! - optimize the code for drawing equal lines
-    add getDirectionFactorByVectorComponent
+    add getDirectionFactorByDirectionComponent
 
 *   [#737](https://github.com/worktile/plait/pull/737) [`91d1685c`](https://github.com/worktile/plait/commit/91d1685c62c4ac4b7717da315e3a32a0838f1174) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - isHorizontalSegment renamed to isOverHorizontal
     isVerticalSegment renamed to isOverVertical